pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "travel-ops-env"
version = "0.1.0"
description = "TravelOps — Business-travel execution & disruption-recovery OpenEnv benchmark"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.25",
]

[tool.setuptools.packages.find]
include = ["server*", "models*", "client*"]
